+++ linear regression engine/engine.py
@@ -3,9 +3,9 @@ from matplotlib import pyplot as plt
 import numpy as np
 
 """ Calculates the mean of a list"""
-def mean(values: List[int]) -> float:
-    sum: int = 0
-    n: int = len(values)
+def mean(values: List[float]) -> float:
+    sum: float = 0
+    n: float = len(values)
 
     for i in range(0, n):
         sum += values[i]
@@ -13,9 +13,9 @@ def mean(values: List[int]) -> float:
     return (sum / n)
 
 """ Calculates the summary statistic for either Sxx or Syy"""
-def ss1(values: List[int]) -> float:
-    sum: int = 0
-    n: int = len(values)
+def ss1(values: List[float]) -> float:
+    sum: float = 0
+    n: float = len(values)
 
     for i in range(0, n):
         sum += (values[i] - mean(values))**2
@@ -23,9 +23,9 @@ def ss1(values: List[int]) -> float:
     return sum
 
 """ Calculates the summary statistic for Sxy"""
-def ss2(xValues: List[int], yValues: List[int]) -> float:
-    sum: int = 0
-    n: int = len(xValues)
+def ss2(xValues: List[float], yValues: List[float]) -> float:
+    sum: float = 0
+    n: float = len(xValues)
 
     for i in range(0, n):
         sum += (xValues[i] - mean(xValues))*(yValues[i] - mean(yValues))
@@ -33,7 +33,7 @@ def ss2(xValues: List[int], yValues: List[int]) -> float:
     return sum
 
 """ Regresion line in the form y = a + bx"""
-def regression_line(xValues: List[int], yValues: List[int]):
+def regression_line(xValues: List[float], yValues: List[float]):
     b = ss2(xValues, yValues) / ss1(xValues)
     a = mean(yValues) - b * mean(xValues)
